Add ShareViewController to handle shared file views
